+++ package.json
@@ -10,6 +10,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "^4.0.3",
+    "sass": "^1.52.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
