--- src/App.css
@@ -1,256 +1,0 @@
-/* Common */
-
-*{
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-}
-
-body {
-  background: #f7f7f7;
-}
-
-ul, nav{
-	list-style: none;
-}
-
-a{
-	text-decoration: none;
-	color: inherit;
-	cursor: pointer;
-
-	opacity: 0.9;
-}
-
-a:hover{
-	color: yellow;
-}
-
-a.btn{
-	color: #fff;
-	border-radius: 4px;
-	text-transform: uppercase;
-	background-color: #2196F3;
-	font-weight: 800;
-	text-align: center;
-}
-
-hr{
-	width: 150px;
-	height: 2px;
-	background-color: #2196F3;
-	border: 0;
-	margin-bottom: 80px;
-}
-
-section{
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	padding: 90px 20px 20px 20px
-}
-
-@media (max-width: 1000px){
-
-	section{
-		padding: 100px 20px 15px 20px;
-	}
-
-}
-
-@media (max-width: 600px){
-
-	section{
-		padding: 130px 20px 20px 20px;
-	}
-
-}
-
-ul.grid{
-	width: 100%;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-
-.section-wrapper{
-	position: relative;
-	justify-content: center;
-	color: #232323;
-	text-align: center;
-  text-transform: capitalize;
-	padding: 50px 20px 15px 20px;
-}
-
-.section-wrapper>h1{
-	font-size: 45px;
-}
-
-.section-wrapper>h3{
-	font-size: 25px;
-}
-/* Header */
-
-header{
-	position: absolute;
-	top: 0;
-	left: 0;
-	z-index: 10;
-	width: 100%;
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	color: #232323;
-	padding: 35px 100px 0;
-}
-
-header h2{
-	font-family: 'Quicksand', sans-serif;
-}
-
-header nav{
-	display: flex;
-}
-
-header nav li{
-	margin: 0 15px;
-}
-
-header nav li:first-child{
-	margin-left: 0;	
-}
-
-header nav li:last-child{
-	margin-right: 0;	
-}
-
-@media (max-width: 1000px){
-	header{
-		padding: 20px 50px;
-	}
-}
-
-@media (max-width: 700px){
-	header{
-		flex-direction: column;		
-	}
-
-	header h2{
-		margin-bottom: 15px;
-	}
-
-	.section-wrapper{
-		padding: 10px 20px 15px 20px;
-	}
-}
-
-.logo{
-  width: 50px;
-  margin-bottom: 10px;
-}
-
-.search-input{
-	font-size: 18px;
-  padding: 3px 5px 3px 5px;
-	color: #555;
-}
-
-/* Title content */
-
-.title-content{
-	position: relative;
-	justify-content: center;
-	color: #232323;
-	text-align: center;
-  text-transform: capitalize;
-}
-
-.title-content>h1{
-	font-size: 45px;
-}
-
-.title-content>h3{
-	font-size: 25px;
-}
-
-/* Categories */
-
-.category { 
-  padding: .5vw;
-  display: -ms-flexbox;
-  -webkit-flex-flow: row wrap; 
-  flex-flow: row wrap; 
-  display: -webkit-box;
-  display: flex;
-}
-
-.category div { 
-  -webkit-box-flex: auto;
-  -ms-flex: auto;
-  flex: auto; 
-  width: 200px; 
-  margin: .5vw; 
-}
-
-.category div img { 
-  width: 100%; 
-  height: auto; 
-}
-
-.mb-15{
-	margin-bottom: 15px;
-}
-
-@media screen and (max-width: 400px) {
-  .category div { margin: 0; }
-  .category { padding: 0; }
-}
-
-/* Footer */
-
-footer{
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	text-align: center;
-	color: #fff;
-	background-color: #232323;
-	padding: 60px 0;
-}
-
-footer ul{
-	display: flex;
-	margin-bottom: 25px;
-	font-size: 32px;
-}
-
-footer ul li{
-	margin: 0 8px;	
-}
-
-footer ul li:first-child{
-	margin-left: 0;	
-}
-
-footer ul li:last-child{
-	margin-right: 0;	
-}
-
-footer p{
-	text-transform: uppercase;
-	font-size: 14px;
-	color: rgba(255,255,255,0.6);
-	margin-bottom: 10px;
-}
-
-footer p a{
-	color: #fff;
-}
-
-@media (max-width: 700px){
-
-	footer{
-		padding: 80px 15px;
-	}
-
-}
