+++ src/App.scss
@@ -1,0 +1,103 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  background: #f7f7f7;
+}
+
+ul,
+nav {
+  list-style: none;
+}
+
+a {
+  text-decoration: none;
+  color: inherit;
+  cursor: pointer;
+  opacity: 0.9;
+}
+
+a:hover {
+  color: yellow;
+}
+
+a.btn {
+  color: #fff;
+  border-radius: 4px;
+  text-transform: uppercase;
+  background-color: #2196F3;
+  font-weight: 800;
+  text-align: center;
+}
+
+hr {
+  width: 150px;
+  height: 2px;
+  background-color: #2196F3;
+  border: 0;
+  margin-bottom: 80px;
+}
+
+section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 90px 20px 20px
+}
+
+@media(max-width: 1000px) {
+  section {
+    padding: 100px 20px 15px;
+  }
+}
+
+@media(max-width: 600px) {
+  section {
+    padding: 130px 20px 20px;
+  }
+}
+
+ul.grid {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.section-wrapper {
+  position: relative;
+  justify-content: center;
+  color: #232323;
+  text-align: center;
+  text-transform: capitalize;
+  padding: 50px 20px 15px;
+}
+
+.section-wrapper > h1 {
+  font-size: 45px;
+}
+
+.section-wrapper > h3 {
+  font-size: 25px;
+}
+
+/* Title content */
+
+.title-content {
+  position: relative;
+  justify-content: center;
+  color: #232323;
+  text-align: center;
+  text-transform: capitalize;
+}
+
+.title-content > h1 {
+  font-size: 45px;
+}
+
+.title-content > h3 {
+  font-size: 25px;
+}
