+++ src/components/category/category.scss
@@ -1,0 +1,34 @@
+.category {
+  padding: 0.5vw;
+  display: -ms-flexbox;
+  -webkit-flex-flow: row wrap;
+  flex-flow: row wrap;
+  display: -webkit-box;
+  display: flex;
+}
+
+.category div {
+  -webkit-box-flex: auto;
+  -ms-flex: auto;
+  flex: auto;
+  width: 200px;
+  margin: 0.5vw;
+}
+
+.category div img {
+  width: 100%;
+  height: auto;
+}
+
+.mb-15 {
+  margin-bottom: 15px;
+}
+
+@media screen and(max-width: 400px) {
+  .category div {
+    margin: 0;
+  }
+  .category {
+    padding: 0;
+  }
+}
