--- src/components/footer.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
-const footer = () => {
-  return (
-    <>
-      <footer>
-        <p>
-          Ecommerce created during Wizeline’s Academy React Bootcamp
-        </p>
-      </footer>
-    </>
-  )
-}
-
-export default footer
