+++ src/components/footer/footer.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import './footer.scss';
+
+const footer = () => {
+  return (
+    <footer>
+      <p>
+        Ecommerce created during Wizeline’s Academy React Bootcamp
+      </p>
+    </footer>
+  )
+}
+
+export default footer
