+++ src/components/header/header.scss
@@ -1,0 +1,63 @@
+header {
+  position: absolute;
+  top: 0;
+  left: 0;
+  z-index: 10;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  color: #232323;
+  padding: 35px 100px 0;
+}
+
+header h2 {
+  font-family: 'Quicksand', sans-serif;
+}
+
+header nav {
+  display: flex;
+}
+
+header nav li {
+  margin: 0 15px;
+}
+
+header nav li:first-child {
+  margin-left: 0;
+}
+
+header nav li:last-child {
+  margin-right: 0;
+}
+
+@media(max-width: 1000px) {
+  header {
+    padding: 20px 50px;
+  }
+}
+
+@media(max-width: 700px) {
+  header {
+    flex-direction: column;
+  }
+
+  header h2 {
+    margin-bottom: 15px;
+  }
+
+  .section-wrapper {
+    padding: 10px 20px 15px;
+  }
+}
+
+.logo {
+  width: 50px;
+  margin-bottom: 10px;
+}
+
+.search-input {
+  font-size: 18px;
+  padding: 3px 5px;
+  color: #555;
+}
