+++ src/components/slider/slider.scss
