+++ src/pages/Home.js
@@ -1,0 +1,16 @@
+import Header from "../components/header/header";
+import Content from "../components/content/content";
+import Footer from "../components/footer/footer";
+
+function Home() {
+	
+	return (
+		<>
+			<Header />
+      <Content />
+			<Footer />
+		</>
+	);
+}
+
+export default Home;
