--- src/pages/Main.js
@@ -1,16 +1,0 @@
-import Header from "../components/header";
-import Content from "../components/content";
-import Footer from "../components/footer";
-
-function Main() {
-	
-	return (
-		<div className="App">
-			<Header />
-      <Content />
-			<Footer />
-		</div>
-	);
-}
-
-export default Main;
